Updated documentation for IExtendedCloudIdentityProvider, with the exception of List Roles

+++ src/corelib/Providers/Rackspace/CloudIdentityProvider.cs
@@ -61,14 +61,26 @@ namespace net.openstack.Providers.Rackspace
             return provider.ListRoles(serviceId, markerId, limit, identity);
         }
 
-        public Role AddRole(string name, string descritpion, CloudIdentity identity)
+        /// <inheritdoc/>
+        public Role AddRole(string name, string description, CloudIdentity identity)
         {
+            if (name == null)
+                throw new ArgumentNullException("name");
+            if (string.IsNullOrEmpty(name))
+                throw new ArgumentException("name cannot be empty");
+
             var provider = GetProvider(identity);
-            return provider.AddRole(name, descritpion, identity);
+            return provider.AddRole(name, description, identity);
         }
 
+        /// <inheritdoc/>
         public Role GetRole(string roleId, CloudIdentity identity)
         {
+            if (roleId == null)
+                throw new ArgumentNullException("roleId");
+            if (string.IsNullOrEmpty(roleId))
+                throw new ArgumentException("roleId cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.GetRole(roleId, identity);
         }
@@ -111,8 +123,18 @@ namespace net.openstack.Providers.Rackspace
             return provider.Authenticate(identity);
         }
 
+        /// <inheritdoc/>
         public bool AddRoleToUser(string userId, string roleId, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (roleId == null)
+                throw new ArgumentNullException("roleId");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(roleId))
+                throw new ArgumentException("roleId cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.AddRoleToUser(userId, roleId, identity: identity);
         }
@@ -165,32 +187,94 @@ namespace net.openstack.Providers.Rackspace
             return provider.DeleteUser(userId, identity: identity);
         }
 
+        /// <inheritdoc/>
         public bool SetUserPassword(string userId, string password, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (password == null)
+                throw new ArgumentNullException("password");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(password))
+                throw new ArgumentException("password cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.SetUserPassword(userId, password, identity: identity);
         }
 
+        /// <inheritdoc/>
         public bool SetUserPassword(User user, string password, CloudIdentity identity)
         {
+            if (user == null)
+                throw new ArgumentNullException("user");
+            if (password == null)
+                throw new ArgumentNullException("password");
+            if (string.IsNullOrEmpty(password))
+                throw new ArgumentException("password cannot be empty");
+            if (string.IsNullOrEmpty(user.Id))
+                throw new ArgumentException("user.Id cannot be null or empty");
+            if (string.IsNullOrEmpty(user.Username))
+                throw new ArgumentException("user.Username cannot be null or empty");
+
             var provider = GetProvider(identity);
             return provider.SetUserPassword(user, password, identity: identity);
         }
 
+        /// <inheritdoc/>
         public bool SetUserPassword(string userId, string username, string password, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (username == null)
+                throw new ArgumentNullException("username");
+            if (password == null)
+                throw new ArgumentNullException("password");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(username))
+                throw new ArgumentException("username cannot be empty");
+            if (string.IsNullOrEmpty(password))
+                throw new ArgumentException("password cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.SetUserPassword(userId, username, password, identity: identity);
         }
 
+        /// <inheritdoc/>
         public UserCredential UpdateUserCredentials(User user, string apiKey, CloudIdentity identity)
         {
+            if (user == null)
+                throw new ArgumentNullException("user");
+            if (apiKey == null)
+                throw new ArgumentNullException("apiKey");
+            if (string.IsNullOrEmpty(apiKey))
+                throw new ArgumentException("apiKey cannot be empty");
+            if (string.IsNullOrEmpty(user.Id))
+                throw new ArgumentException("user.Id cannot be null or empty");
+            if (string.IsNullOrEmpty(user.Username))
+                throw new ArgumentException("user.Username cannot be null or empty");
+
             var provider = GetProvider(identity);
             return provider.UpdateUserCredentials(user, apiKey, identity: identity);
         }
 
+        /// <inheritdoc/>
         public UserCredential UpdateUserCredentials(string userId, string username, string apiKey, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (username == null)
+                throw new ArgumentNullException("username");
+            if (apiKey == null)
+                throw new ArgumentNullException("apiKey");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(username))
+                throw new ArgumentException("username cannot be empty");
+            if (string.IsNullOrEmpty(apiKey))
+                throw new ArgumentException("apiKey cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.UpdateUserCredentials(userId, username, apiKey, identity: identity);
         }
@@ -207,14 +291,30 @@ namespace net.openstack.Providers.Rackspace
             return provider.ListUserCredentials(userId, identity: identity);
         }
 
+        /// <inheritdoc/>
         public UserCredential UpdateUserCredentials(string userId, string apiKey, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (apiKey == null)
+                throw new ArgumentNullException("apiKey");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(apiKey))
+                throw new ArgumentException("apiKey cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.UpdateUserCredentials(userId, apiKey, identity: identity);
         }
 
+        /// <inheritdoc/>
         public bool DeleteUserCredentials(string userId, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.DeleteUserCredentials(userId, identity: identity);
         }
@@ -259,8 +359,18 @@ namespace net.openstack.Providers.Rackspace
             return provider.GetToken(identity, forceCacheRefresh);
         }
 
+        /// <inheritdoc/>
         public bool DeleteRoleFromUser(string userId, string roleId, CloudIdentity identity)
         {
+            if (userId == null)
+                throw new ArgumentNullException("userId");
+            if (roleId == null)
+                throw new ArgumentNullException("roleId");
+            if (string.IsNullOrEmpty(userId))
+                throw new ArgumentException("userId cannot be empty");
+            if (string.IsNullOrEmpty(roleId))
+                throw new ArgumentException("roleId cannot be empty");
+
             var provider = GetProvider(identity);
             return provider.DeleteRoleFromUser(userId, roleId, identity: identity);
         }
